+++ walletkit-core/src/credential_storage/platform/ios/blob_store.rs
@@ -32,9 +32,7 @@ fn io_error<S: Into<String>>(context: S, err: std::io::Error) -> StorageError {
     }
 }
 
-// =============================================================================
 // IosBlobStore
-// =============================================================================
 
 /// iOS file system-based implementation of `AtomicBlobStore`.
 ///
@@ -259,136 +257,5 @@ impl AtomicBlobStore for IosBlobStore {
     fn exists(&self, filename: &str) -> StorageResult<bool> {
         let path = self.blob_path(filename);
         Ok(path.exists())
-    }
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    fn create_temp_store() -> (TempDir, IosBlobStore) {
-        let temp_dir = TempDir::new().unwrap();
-        let store = IosBlobStore::new(temp_dir.path()).unwrap();
-        (temp_dir, store)
-    }
-
-    #[test]
-    fn test_read_nonexistent() {
-        let (_temp, store) = create_temp_store();
-        let result = store.read("nonexistent.bin").unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_write_and_read() {
-        let (_temp, store) = create_temp_store();
-
-        let data = b"Hello, World!";
-        store.write_atomic("test.bin", data).unwrap();
-
-        let read_data = store.read("test.bin").unwrap().unwrap();
-        assert_eq!(read_data, data);
-    }
-
-    #[test]
-    fn test_overwrite() {
-        let (_temp, store) = create_temp_store();
-
-        store.write_atomic("test.bin", b"first").unwrap();
-        store.write_atomic("test.bin", b"second").unwrap();
-
-        let data = store.read("test.bin").unwrap().unwrap();
-        assert_eq!(data, b"second");
-    }
-
-    #[test]
-    fn test_delete() {
-        let (_temp, store) = create_temp_store();
-
-        store.write_atomic("test.bin", b"data").unwrap();
-        assert!(store.exists("test.bin").unwrap());
-
-        store.delete("test.bin").unwrap();
-        assert!(!store.exists("test.bin").unwrap());
-    }
-
-    #[test]
-    fn test_delete_nonexistent() {
-        let (_temp, store) = create_temp_store();
-
-        // Should not error
-        store.delete("nonexistent.bin").unwrap();
-    }
-
-    #[test]
-    fn test_exists() {
-        let (_temp, store) = create_temp_store();
-
-        assert!(!store.exists("test.bin").unwrap());
-
-        store.write_atomic("test.bin", b"data").unwrap();
-
-        assert!(store.exists("test.bin").unwrap());
-    }
-
-    #[test]
-    fn test_empty_data() {
-        let (_temp, store) = create_temp_store();
-
-        store.write_atomic("empty.bin", b"").unwrap();
-
-        let data = store.read("empty.bin").unwrap().unwrap();
-        assert!(data.is_empty());
-    }
-
-    #[test]
-    fn test_large_data() {
-        let (_temp, store) = create_temp_store();
-
-        let data = vec![0xABu8; 1024 * 1024]; // 1 MB
-        store.write_atomic("large.bin", &data).unwrap();
-
-        let read_data = store.read("large.bin").unwrap().unwrap();
-        assert_eq!(read_data, data);
-    }
-
-    #[test]
-    fn test_special_characters_in_filename() {
-        let (_temp, store) = create_temp_store();
-
-        // Note: Some special characters may not work on all file systems
-        let filename = "test-file_123.bin";
-        store.write_atomic(filename, b"data").unwrap();
-
-        let data = store.read(filename).unwrap().unwrap();
-        assert_eq!(data, b"data");
-    }
-
-    #[test]
-    fn test_no_temp_file_left_on_success() {
-        let (temp, store) = create_temp_store();
-
-        store.write_atomic("test.bin", b"data").unwrap();
-
-        // Check that no temp file exists
-        let temp_path = temp.path().join(".test.bin.tmp");
-        assert!(!temp_path.exists());
-    }
-
-    #[test]
-    fn test_creates_directory() {
-        let temp_dir = TempDir::new().unwrap();
-        let nested_path = temp_dir.path().join("nested").join("dir");
-
-        let store = IosBlobStore::new(&nested_path).unwrap();
-        assert!(nested_path.exists());
-
-        store.write_atomic("test.bin", b"data").unwrap();
-        assert!(store.read("test.bin").unwrap().is_some());
     }
 }
