+++ walletkit-core/src/credential_storage/platform/ios/vault_store.rs
@@ -30,9 +30,7 @@ fn io_error<S: Into<String>>(context: S, err: std::io::Error) -> StorageError {
     }
 }
 
-// =============================================================================
 // IosVaultStore
-// =============================================================================
 
 /// iOS file-based implementation of `VaultFileStore`.
 ///
@@ -370,202 +368,5 @@ impl IosVaultStore {
     #[must_use]
     pub fn exists(&self) -> bool {
         self.path.exists()
-    }
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    fn create_temp_store() -> (TempDir, IosVaultStore) {
-        let temp_dir = TempDir::new().unwrap();
-        let vault_path = temp_dir.path().join("account.vault");
-        let store = IosVaultStore::new(&vault_path).unwrap();
-        (temp_dir, store)
-    }
-
-    #[test]
-    fn test_create_and_open() {
-        let (_temp, store) = create_temp_store();
-
-        // File doesn't exist yet
-        assert!(!store.exists());
-
-        // Create it
-        store.create_new().unwrap();
-        assert!(store.exists());
-
-        // Should be empty
-        assert_eq!(store.len().unwrap(), 0);
-        assert!(store.is_empty().unwrap());
-    }
-
-    #[test]
-    fn test_create_fails_if_exists() {
-        let (_temp, store) = create_temp_store();
-
-        store.create_new().unwrap();
-
-        // Creating again should fail
-        let result = store.create_new();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_open_fails_if_not_exists() {
-        let (_temp, store) = create_temp_store();
-
-        let result = store.open_existing();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_write_and_read_at() {
-        let (_temp, store) = create_temp_store();
-        store.create_new().unwrap();
-
-        let data = b"Hello, World!";
-        store.write_at(0, data).unwrap();
-
-        let buf = store.read_at(0, data.len() as u32).unwrap();
-
-        assert_eq!(buf, data);
-    }
-
-    #[test]
-    fn test_append() {
-        let (_temp, store) = create_temp_store();
-        store.create_new().unwrap();
-
-        let offset1 = store.append(b"first").unwrap();
-        assert_eq!(offset1, 0);
-
-        let offset2 = store.append(b"second").unwrap();
-        assert_eq!(offset2, 5);
-
-        assert_eq!(store.len().unwrap(), 11);
-    }
-
-    #[test]
-    fn test_write_at_extends_file() {
-        let (_temp, store) = create_temp_store();
-        store.create_new().unwrap();
-
-        // Write at offset 100 (beyond current end)
-        store.write_at(100, b"data").unwrap();
-
-        assert_eq!(store.len().unwrap(), 104);
-
-        // Read back
-        let buf = store.read_at(100, 4).unwrap();
-        assert_eq!(buf, b"data");
-    }
-
-    #[test]
-    fn test_set_len_truncate() {
-        let (_temp, store) = create_temp_store();
-        store.create_new().unwrap();
-
-        store.write_at(0, b"Hello, World!").unwrap();
-        assert_eq!(store.len().unwrap(), 13);
-
-        store.set_len(5).unwrap();
-        assert_eq!(store.len().unwrap(), 5);
-
-        let buf = store.read_at(0, 5).unwrap();
-        assert_eq!(buf, b"Hello");
-    }
-
-    #[test]
-    fn test_set_len_extend() {
-        let (_temp, store) = create_temp_store();
-        store.create_new().unwrap();
-
-        store.set_len(100).unwrap();
-        assert_eq!(store.len().unwrap(), 100);
-    }
-
-    #[test]
-    fn test_sync() {
-        let (_temp, store) = create_temp_store();
-        store.create_new().unwrap();
-
-        store.write_at(0, b"data").unwrap();
-        store.sync().unwrap(); // Should not error
-    }
-
-    #[test]
-    fn test_empty_operations() {
-        let (_temp, store) = create_temp_store();
-        store.create_new().unwrap();
-
-        // Empty write should succeed
-        store.write_at(0, b"").unwrap();
-        assert_eq!(store.len().unwrap(), 0);
-
-        // Empty read should succeed
-        let buf = store.read_at(0, 0).unwrap();
-        assert!(buf.is_empty());
-
-        // Empty append should return current length
-        let offset = store.append(b"").unwrap();
-        assert_eq!(offset, 0);
-    }
-
-    #[test]
-    fn test_reopen() {
-        let (temp, store) = create_temp_store();
-
-        // Create and write
-        store.create_new().unwrap();
-        store.write_at(0, b"persisted data").unwrap();
-        store.sync().unwrap();
-
-        // Drop the store
-        drop(store);
-
-        // Create new store for same path
-        let vault_path = temp.path().join("account.vault");
-        let store2 = IosVaultStore::new(&vault_path).unwrap();
-
-        // Open existing file
-        store2.open_existing().unwrap();
-
-        // Read back data
-        let buf = store2.read_at(0, 14).unwrap();
-        assert_eq!(buf, b"persisted data");
-    }
-
-    #[test]
-    fn test_creates_parent_directory() {
-        let temp_dir = TempDir::new().unwrap();
-        let vault_path = temp_dir.path().join("nested").join("dir").join("account.vault");
-
-        let store = IosVaultStore::new(&vault_path).unwrap();
-        store.create_new().unwrap();
-
-        assert!(vault_path.exists());
-    }
-
-    #[test]
-    fn test_open_or_create() {
-        let (_temp, store) = create_temp_store();
-
-        // File doesn't exist, should create
-        store.open_or_create().unwrap();
-        assert!(store.exists());
-
-        // Write some data
-        store.write_at(0, b"test").unwrap();
-        store.sync().unwrap();
-
-        // Verify the store was created and is functional
-        let data = store.read_at(0, 4).unwrap();
-        assert_eq!(data, b"test");
     }
 }
